address Himanshu's comments
